Color Schemes thru ScriptableObjects
Can Create TypeMaterials color schemes that assign specific materials to specific TileTypes. This can be swapped out per level, and a default scheme can be assigned to the game manager.

DiceMappings assets can be created that map which side of the dice is which TileType.

On Level Load, the Dice sides are assigned and colors are updated